# docs/swagger.yaml
definitions:
  domain.CreateEventPayload:
    properties:
      author:
        type: string
      capacity:
        type: integer
      date:
        type: string
      description:
        type: string
      duration:
        type: string
      file_name:
        type: string
      is_online:
        type: string
      location:
        type: string
      price:
        type: number
      registration_link:
        type: string
      reservation_start_date:
        type: string
      reseveration_end_date:
        type: string
      slug:
        type: string
      speakers:
        items:
          type: string
        type: array
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
    required:
    - author
    - capacity
    - date
    - description
    - duration
    - file_name
    - is_online
    - location
    - slug
    - status
    - title
    - type
    type: object
  domain.Event:
    properties:
      author:
        type: string
      capacity:
        type: integer
      created_at:
        type: string
      created_by:
        type: integer
      date:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: integer
      description:
        type: string
      duration:
        type: string
      id:
        type: integer
      image_event:
        type: string
      location:
        type: string
      price:
        description: 0 == free
        type: number
      registration_link:
        type: string
      reservation_start_date:
        type: string
      reseveration_end_date:
        type: string
      slug:
        type: string
      speakers:
        description: Ensure foreign key is correctly referenced
        items:
          $ref: '#/definitions/domain.EventSpeaker'
        type: array
      status:
        description: comming soon
        type: string
      tags:
        description: Ensure foreign key is correctly referenced
        items:
          $ref: '#/definitions/domain.EventTag'
        type: array
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
...
  domain.EventPayPayload:
    properties:
      image_proof_payment:
        type: string
      net_amount:
        type: number
      order_no:
        type: string
    type: object
  domain.EventSpeaker:
    properties:
      event_id:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  domain.EventTag:
    properties:
      event_id:
        type: integer
      id:
        type: integer
      tags:
        type: string
    type: object
  domain.HttpResponse:
    properties:
      code:
        type: integer
      data: {}
      limit:
        type: integer
      message:
        type: string
      page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  domain.RegisterEventPayload:
    properties:
      email:
        type: string
      event_id:
        type: integer
      image_proof_payment:
        type: string
      name:
        type: string
      net_amount:
        type: number
      phone_number:
        type: string
      user_id:
        type: string
    type: object
...
  domain.RegistrationEvent:
    properties:
      created_at:
        type: string
      created_by_user_id:
        type: integer
      deleted_at:
        type: string
      deleted_by_user_id:
        type: integer
      email:
        type: string
      event_id:
        description: lock event
        type: integer
      id:
        type: integer
      image_proof_payment:
        type: string
      name:
        type: string
      order_no:
        type: string
      payment_date:
        type: string
      phone_number:
        type: string
      status:
        description: register, pay, approve/cancel/decline
        type: string
      up_to_you:
        type: string
      updated_at:
        type: string
      updated_by_user_id:
        type: integer
      user_id:
        description: lock user
        type: string
    type: object
  domain.UpdateEventPayload:
    properties:
      author:
        type: string
      capacity:
        type: integer
      date:
        type: string
      description:
        type: string
      duration:
        type: string
      file_name:
        type: string
      is_online:
        type: string
      location:
        type: string
      price:
        type: number
      registration_link:
        type: string
      reservation_start_date:
        type: string
      reseveration_end_date:
        type: string
      slug:
        type: string
      speakers:
        items:
          type: string
        type: array
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
    required:
    - author
    - capacity
    - date
    - description
    - duration
    - file_name
    - is_online
    - location
    - slug
    - status
    - title
    - type
    type: object
...
info:
  contact:
    email: hammercode.org
    name: API Support
    url: http://hammercode.org/support
  description: This is the API documentation for this service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Golang API (LMS-BE)
  version: "1.0"
paths:
  /api/v1/admin/events:
    get:
      consumes:
      - application/json
      description: This endpoint use to get events by filter
      parameters:
      - description: string
        in: query
        name: limit
        required: true
        type: string
      - description: string
        in: query
        name: page
        required: true
        type: string
      - description: string
        in: query
        name: orderBy
        type: string
      - description: string
        in: query
        name: start_date
        type: string
      - description: string
        in: query
        name: end_date
        type: string
      - description: string
        in: query
        name: title
        type: string
      - description: string
        in: query
        name: type
        type: string
      - description: string
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Get Events
      tags:
      - Event
    post:
      consumes:
      - application/json
      description: This endpoint create event
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateEventPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      security:
      - BearerAuth: []
      summary: Create Event
      tags:
      - Event
  /api/v1/admin/events/{id}:
    delete:
      consumes:
      - application/json
      description: This endpoint use to delete event by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      security:
      - BearerAuth: []
      summary: Delete Event By Id
      tags:
      - Event
    get:
      consumes:
      - application/json
      description: This endpoint use to get detail event by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Get Detail Event by ID
      tags:
      - Event
    patch:
      consumes:
      - application/json
      description: This endpoint update event
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateEventPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      security:
      - BearerAuth: []
      summary: Update Event
      tags:
      - Event
...
  /api/v1/events:
    get:
      consumes:
      - application/json
      description: This endpoint use to get events by filter
      parameters:
      - description: string
        in: query
        name: limit
        required: true
        type: string
      - description: string
        in: query
        name: page
        required: true
        type: string
      - description: string
        in: query
        name: orderBy
        type: string
      - description: string
        in: query
        name: start_date
        type: string
      - description: string
        in: query
        name: end_date
        type: string
      - description: string
        in: query
        name: title
        type: string
      - description: string
        in: query
        name: type
        type: string
      - description: string
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Get List Available Event
      tags:
      - Event
  /api/v1/events/{id}:
    get:
      consumes:
      - application/json
      description: This endpoint use to get event by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Get Detail Event by ID
      tags:
      - Event
  /api/v1/events/pay:
    post:
      consumes:
      - application/json
      description: This endpoint pay the event
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.EventPayPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Pay Event
      tags:
      - Event
...
  /api/v1/events/registrations:
    get:
      consumes:
      - application/json
      description: This endpoint use to get list of user that register on event
      parameters:
      - description: string
        in: query
        name: limit
        required: true
        type: string
      - description: string
        in: query
        name: page
        required: true
        type: string
      - description: string
        in: query
        name: event_id
        type: string
      - description: string
        in: query
        name: start_date
        type: string
      - description: string
        in: query
        name: end_date
        type: string
      - description: string
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.RegistrationEvent'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Get List Register user to event
      tags:
      - Event
    post:
      consumes:
      - application/json
      description: This endpoint use to register event
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterEventPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Register Event
      tags:
      - Event
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"